+++ crates/bcr-wallet-core/src/pocket/debit.rs
@@ -77,6 +77,7 @@ impl Pocket {
         };
         Ok(active_info.clone())
     }
+
     async fn find_active_keyset(
         &self,
         keysets_info: &[KeySetInfo],
@@ -294,11 +295,12 @@ impl wallet::Pocket for Pocket {
 impl wallet::DebitPocket for Pocket {
     async fn reclaim_proofs(
         &self,
+        ys: &[cdk01::PublicKey],
         keysets_info: &[KeySetInfo],
         client: &dyn MintConnector,
         safe_mode: wallet::SafeMode,
     ) -> Result<Amount> {
-        let pendings = self.pdb.list_pending().await?;
+        let pendings = self.pdb.load_proofs(ys).await?;
         let pendings_len = pendings.len();
         let active_keys = self.find_active_keyset(keysets_info, client).await?;
         let (reclaimed, _) = self
@@ -460,7 +462,10 @@ impl wallet::DebitPocket for Pocket {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{utils::tests::MockMintConnector, wallet::Pocket};
+    use crate::{
+        utils::tests::MockMintConnector,
+        wallet::{DebitPocket, Pocket},
+    };
     use bcr_common::core_tests;
     use mockall::predicate::*;
 
@@ -469,6 +474,7 @@ mod tests {
         let seed = [0u8; 64];
         super::Pocket::new(unit, pdb, mdb, seed)
     }
+
     #[tokio::test]
     async fn debit_receive_proofs() {
         let (info, keyset) = core_tests::generate_random_ecash_keyset();
@@ -517,5 +523,70 @@ mod tests {
             .await
             .unwrap();
         assert_eq!(cashed, Amount::from(24u64));
+    }
+
+    #[tokio::test]
+    async fn debit_reclaim_proofs() {
+        let (info, keyset) = core_tests::generate_random_ecash_keyset();
+        let kid = info.id;
+        let k_infos = vec![KeySetInfo::from(info)];
+        let amounts = [Amount::from(8u64), Amount::from(16u64)];
+        let proofs = core_tests::generate_random_ecash_proofs(&keyset, &amounts);
+
+        let ys: Vec<cdk01::PublicKey> = proofs.iter().map(|p| p.y().expect("valid y")).collect();
+
+        let mdb = MockMintMeltRepository::new();
+        let mut pdb = MockPocketRepository::new();
+        let mut connector = MockMintConnector::new();
+        let cloned_keyset = keyset.clone();
+
+        connector
+            .expect_get_mint_keyset()
+            .times(1)
+            .with(eq(kid))
+            .returning(move |_| Ok(KeySet::from(cloned_keyset.clone())));
+        let proofs_clone = proofs.clone();
+        let ys_clone = ys.clone();
+        pdb.expect_load_proofs()
+            .times(1)
+            .with(eq(ys_clone))
+            .returning(move |_| {
+                let mut map = HashMap::new();
+                map.insert(proofs_clone[0].y().unwrap(), proofs_clone[0].clone());
+                map.insert(proofs_clone[1].y().unwrap(), proofs_clone[1].clone());
+                Ok(map)
+            });
+        pdb.expect_counter()
+            .times(1)
+            .with(eq(kid))
+            .returning(|_| Ok(0));
+        pdb.expect_increment_counter()
+            .times(1)
+            .with(eq(kid), eq(0), eq(2))
+            .returning(|_, _, _| Ok(()));
+        connector
+            .expect_post_swap()
+            .times(1)
+            .returning(move |request| {
+                let amounts = request
+                    .outputs()
+                    .iter()
+                    .map(|b| b.amount)
+                    .collect::<Vec<_>>();
+                let signatures = core_tests::generate_ecash_signatures(&keyset, &amounts);
+                let response = cdk03::SwapResponse { signatures };
+                Ok(response)
+            });
+        pdb.expect_store_new().times(2).returning(|p| {
+            let y = p.y().expect("Hash to curve should not fail");
+            Ok(y)
+        });
+
+        let pocket = pocket(Arc::new(pdb), Arc::new(mdb));
+        let reclaimed = pocket
+            .reclaim_proofs(&ys, &k_infos, &connector, SafeMode::Disabled)
+            .await
+            .expect("reclaim works");
+        assert_eq!(reclaimed, Amount::from(24u64));
     }
 }